worker/requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
requests==2.31.0
toml==0.10.2
cryptography==41.0.7
